feat(test): harmonize ContractDefinitionStore test to use a common test specification
