+++ core/control-plane/control-plane-core/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     api(project(":core:control-plane:transfer"))
     implementation(project(":core:common:util"))
     testImplementation(testFixtures(project(":spi:common:core-spi")))
+    testImplementation(testFixtures(project(":spi:control-plane:contract-spi")))
 }
 
 publishing {
