+++ extensions/control-plane/store/sql/contract-definition-store-sql/build.gradle.kts
@@ -29,6 +29,7 @@ dependencies {
 
     testImplementation(project(":extensions:common:junit"))
     testImplementation(project(":core:common:policy-evaluator"))
+    testImplementation(testFixtures(project(":spi:control-plane:contract-spi")))
     testImplementation("org.assertj:assertj-core:${assertj}")
     testImplementation(testFixtures(project(":core:common:util")))
     testImplementation("org.postgresql:postgresql:${postgresVersion}")
