+++ spi/control-plane/contract-spi/build.gradle.kts
@@ -12,14 +12,24 @@
  *
  */
 
+val jupiterVersion: String by project
+val assertj: String by project
+
+
 plugins {
     `java-library`
+    `java-test-fixtures`
 }
 
 
 dependencies {
     api(project(":spi:common:core-spi"))
     api(project(":spi:control-plane:policy-spi"))
+
+    testFixturesImplementation("org.junit.jupiter:junit-jupiter-api:${jupiterVersion}")
+    testFixturesImplementation("org.junit.jupiter:junit-jupiter-params:${jupiterVersion}")
+    testFixturesImplementation("org.assertj:assertj-core:${assertj}")
+
 }
 
 publishing {
